requirements/base.txt
Django>=3.2,<3.3
django-dotenv==1.4.1
wagtail>=2.14,<2.15
wagtailfontawesome>=1.1.3,<1.2
django-debug-toolbar>=3.2,<4
pymysql>=1.1.0